refactor: reescrita do método destroy com validação e tratamento de erros das tarefas

+++ app/Http/Controllers/TarefasController.php
@@ -8,6 +8,8 @@ use Illuminate\Http\Request;
 
 use App\Traits\ApiResponse;
 
+use App\Http\Requests\UpdateTarefaRequest;
+
 
 class TarefasController extends Controller
 {
@@ -27,15 +29,52 @@ class TarefasController extends Controller
     public function show(Tarefa $tarefa){
         return $this->success($tarefa->load('subtarefas'), 'Tarefa encontrada');
     }
-    public function update(Request $request, Tarefa $tarefa){
-        $tarefa->update($request->only(['titulo', 'status']));
-        return $this->success($tarefa, 'Tarefa atualizada com sucesso');
+
+    public function update(UpdateTarefaRequest $request, Tarefa $tarefa){
+        try{
+            $dados = $request->only(['titulo', 'status']);
+            $tarefa->update($dados);
+            return response()->json([
+                'message' => 'Tarefa atualizada com sucesso!',
+                'data' => $tarefa
+            ], 200);
+        } catch (\Exception $e) {
+            return response()->json([
+                'message' => 'Erro ao atualizar a tarefa',
+                'error' => $e->getMessage()
+            ], 500);
+        }
     }
 
-    public function destroy(Tarefa $tarefa){
-        $tarefa->delete();
-        return $this->success(null, 'Tarefa excluida com sucesso', 204);
-    }
+    public function destroy($id){
 
+        $tarefa = Tarefa::find($id);
+
+        if(!$tarefa){
+            return response()->json([
+                'message' => 'Tarefa não encontrada'
+            ], 404);
+        }
+
+        if($tarefa->status === 'concluida'){
+            return response()->json([
+                'message' => 'Tarefas concluídas não podem ser deletadas'
+            ], 403);
+        }
+
+        try{
+            $tarefa->delete();
+
+            return response()->json([
+                'message' => 'Tarefa deletada com sucesso.'
+            ], 200);
+        } catch(\Exception $e){
+            return response()->json([
+                'message' => 'Erro ao deletar tarefa.',
+                'error' => $e->getMessage()
+            ], 500);
+
+        }
+    }
 
 }
